refactor(config): give JSON Patch one implementation
The gateway's `PATCH /api/config` and the CLI's `zeroclaw config patch` each
carried their own op loop over the same five operations. A component test
existed solely to assert the two still produced identical error envelopes --
a test that only needs to exist because the code was duplicated.

`zeroclaw_config::patch` now owns the shape (`PatchOp`, `PatchOpResult`), the
parser, the op loop, and the prop-metadata helpers the loop needs. The
gateway's 183-line inline loop becomes a five-line call; its structs are
re-exported so no caller of `api_config::PatchOp` has to change.

Applying is deliberately separate from persisting. `apply_patch_ops` takes a
`&mut Config` and mutates in memory only; the caller decides whether that
config is saved, swapped into a running gateway, or discarded. That is what
lets a caller apply a patch to a *copy* to answer "what would this do?"
without committing it -- the property an approval preview needs, pinned by
`applying_to_a_clone_leaves_the_original_alone`.

No behaviour change. The CLI is untouched in this commit, so
`config_patch_json_failed_op_matches_http_error_envelope` -- which compares
the CLI's envelope against the gateway's, field by field including the
message -- passing is direct evidence that the extracted implementation is
byte-identical to the one it replaced.

Next: point the CLI at the same code and delete its copy.
