+++ crates/zeroclaw-config/src/lib.rs
@@ -18,6 +18,7 @@ pub mod helpers;
 pub mod migration;
 pub mod multi_agent;
 pub mod pairing;
+pub mod patch;
 pub mod paths;
 pub mod platform;
 pub mod policy;
